JAXB stuff for wrappers XML Schema
Replaced the wrappers written in JAVA to be now auto-generated from
schema file.
XSD files are:
um-blueprint.xsd : It contains the core elements for API
um.xsd : Contains the Transfer Objects which will be exposed by the REST
layer.

A jaxb sample also committed, for help.

Bindings file is not in use as of now.